refactor(pipeline): extract shared helpers, fix type safety, and sync docs
- Extract buildOnFinish to shared module (lib/ai/pipeline/on-finish.ts)
- Add selectedDocumentIds scope filtering in classification pipeline
- Parallelize data fetch and message compression in classification
- Fix TypeScript errors (unknown classification type, optional messages)
- Simplify BuildOnFinishParams (lastMessage -> lastMessageRole)
- Fix p.text nullability inconsistency in cover-letter route
- Remove unused PipelineParams/PipelineResult types
- Sync docs (features, changelog) for pipeline branching
- Update test mocks to include buildOnFinish via importOriginal
