+++ app/api/chat/interview/route.ts
@@ -1,6 +1,5 @@
 import { convertToModelMessages, type UIMessage } from "ai"
 import { NextResponse } from "next/server"
-import { MessageRole } from "@prisma/client"
 import { createClient } from "@/lib/supabase/server"
 import { prisma } from "@/lib/prisma"
 import { getLanguageModel, AiSettingsNotFoundError } from "@/lib/ai/provider"
@@ -10,7 +9,7 @@ import { createReadDocumentTool } from "@/lib/ai/tools"
 import { interviewChatSchema } from "@/lib/validations/interview"
 import { recordUsage } from "@/lib/token-usage/service"
 import { checkQuotaExceeded } from "@/lib/token-usage/quota"
-import { selectPipeline, handleMultiStep, handleClassification, interviewClassificationSchema } from "@/lib/ai/pipeline"
+import { selectPipeline, handleMultiStep, handleClassification, interviewClassificationSchema, buildOnFinish } from "@/lib/ai/pipeline"
 
 export const maxDuration = 120
 
@@ -124,7 +123,7 @@ export async function POST(request: Request) {
     const modelMessages = await convertToModelMessages(messages as UIMessage[])
 
     const onFinish = buildOnFinish({
-      conversationId, lastMessage, lastMessageContent,
+      conversationId, lastMessageRole: lastMessage.role, lastMessageContent,
       userId: user.id, aiProvider, modelId, isServerKey,
       feature: "INTERVIEW",
     })
@@ -185,60 +184,3 @@ export async function POST(request: Request) {
   }
 }
 
-// ─── 헬퍼 ──────────────────────────────────────────────────────────────────
-
-interface BuildOnFinishParams {
-  conversationId: string
-  lastMessage: { role: string; parts?: { type: string; text: string }[]; content?: string }
-  lastMessageContent: string
-  userId: string
-  aiProvider: string
-  modelId: string
-  isServerKey: boolean
-  feature: "COVER_LETTER" | "INTERVIEW"
-}
-
-function buildOnFinish(params: BuildOnFinishParams) {
-  return async ({ text, usage, steps }: {
-    text: string
-    usage?: { inputTokens?: number; outputTokens?: number }
-    steps: { toolCalls?: { toolName: string }[] }[]
-  }) => {
-    const toolCalls = steps.flatMap(s => s.toolCalls ?? [])
-    if (toolCalls.length > 0) {
-      console.log(`[${params.feature.toLowerCase()}] 도구 호출 ${toolCalls.length}건:`, toolCalls.map(tc => tc.toolName).join(", "))
-    } else {
-      console.log(`[${params.feature.toLowerCase()}] 도구 호출 없음`)
-    }
-
-    const ops = [
-      ...(params.lastMessage.role === "user" && params.lastMessageContent
-        ? [prisma.message.create({
-            data: { conversationId: params.conversationId, role: MessageRole.USER, content: params.lastMessageContent },
-          })]
-        : []),
-      ...(text
-        ? [prisma.message.create({
-            data: { conversationId: params.conversationId, role: MessageRole.ASSISTANT, content: text },
-          })]
-        : []),
-    ]
-    if (ops.length > 0) {
-      await prisma.$transaction(ops)
-    }
-
-    if (usage) {
-      await recordUsage({
-        userId: params.userId,
-        provider: params.aiProvider,
-        model: params.modelId,
-        feature: params.feature,
-        promptTokens: usage.inputTokens ?? 0,
-        completionTokens: usage.outputTokens ?? 0,
-        totalTokens: (usage.inputTokens ?? 0) + (usage.outputTokens ?? 0),
-        isServerKey: params.isServerKey,
-        metadata: { conversationId: params.conversationId },
-      }).catch((e) => console.error("토큰 사용량 기록 실패:", e))
-    }
-  }
-}
