+++ docs/features/05-ai-infra.md
@@ -84,6 +84,40 @@ AI SDK v6의 `UIMessage` (parts 배열 기반) 구조에 맞춘 재사용 가능
 | `PUT` | `/api/settings/ai` | AI 설정 저장/업데이트 (upsert) |
 | `POST` | `/api/settings/ai/validate` | API 키 유효성 검증 (제공자별 연결 테스트) |
 
+## 파이프라인 분기 (Pipeline Branching)
+
+AI 제공자에 따라 채팅 처리 방식을 분기한다. `selectPipeline(provider)`로 결정.
+
+| 파이프라인 | 대상 제공자 | 방식 |
+|-----------|-----------|------|
+| **multi-step** | OpenAI | 기존 Tool Use 방식. `streamText` + `tools` + `stopWhen`으로 LLM이 필요시 도구 호출 |
+| **classification** | Anthropic, Google 등 | 분류→데이터 수집→(압축)→응답. 도구 없이 서버에서 데이터를 직접 주입 |
+
+### multi-step 파이프라인
+
+기존 Tool Use 방식을 `handleMultiStep()`으로 래핑. LLM이 `readDocument`, `readCareerNote`, `saveCareerNote` 도구를 자율적으로 호출한다.
+
+### classification 파이프라인
+
+`handleClassification()`이 3단계로 처리:
+
+1. **분류** — `classify()`가 `generateText` + `Output.object`로 사용자 메시지를 분석. 어떤 문서를 읽을지, 커리어노트 비교가 필요한지, 대화 압축이 필요한지 판단
+2. **데이터 수집** — 분류 결과에 따라 서버에서 문서/커리어노트 전문을 DB 조회. 대화가 길면 `compressMessages()`로 이전 대화 요약
+3. **응답 생성** — 수집된 데이터를 시스템 프롬프트에 주입하고 `streamText` (tools 없음)
+
+#### 분류 스키마
+
+- **자소서**: `documentsToRead`, `compareCareerNotes`, `needsCompression`
+- **면접**: `documentsToRead`, `needsCompression`
+
+#### 대화 압축
+
+최근 4턴을 유지하고, 이전 대화를 LLM으로 3~5문장 요약. 압축 실패 시 원본 유지 (graceful fallback).
+
+#### 토큰 사용량
+
+분류·압축 단계의 토큰 사용량은 `preStageUsage`로 별도 반환하여 route handler에서 `recordUsage`에 합산.
+
 ## 주요 파일
 
 | 파일 | 역할 |
@@ -94,6 +128,12 @@ AI SDK v6의 `UIMessage` (parts 배열 기반) 구조에 맞춘 재사용 가능
 | `lib/ai/context.ts` | 요약 기반 컨텍스트 빌더 |
 | `lib/ai/tools/*.ts` | 채팅 도구 (readDocument, readCareerNote, saveCareerNote) |
 | `lib/ai/prompts/*.ts` | 프롬프트 템플릿 (cover-letter, interview, insight-extraction) |
+| `lib/ai/pipeline/index.ts` | `selectPipeline()` + 모듈 re-export |
+| `lib/ai/pipeline/multi-step.ts` | multi-step 파이프라인 (Tool Use 래핑) |
+| `lib/ai/pipeline/classification.ts` | classification 파이프라인 (분류→수집→응답) |
+| `lib/ai/pipeline/classify.ts` | LLM 분류 함수 (`generateText` + `Output.object`) |
+| `lib/ai/pipeline/compress.ts` | 대화 압축 (이전 대화 요약) |
+| `lib/ai/pipeline/schema.ts` | 분류 스키마 (자소서/면접별) |
 | `lib/settings/service.ts` | 설정 CRUD 서비스 레이어 |
 | `app/api/settings/ai/route.ts` | GET, PUT 엔드포인트 |
 | `app/(dashboard)/settings/page.tsx` | 설정 페이지 (RSC) |
